집중중에만 방해 알림 보내도록
